+++ dflib/src/test/java/org/dflib/BooleanSeries_TrueFalseTest.java
@@ -1,6 +1,5 @@
 package org.dflib;
 
-import org.dflib.Series;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -10,7 +9,7 @@ public class BooleanSeries_TrueFalseTest {
 
     @Test
     public void isTrue() {
-        assertTrue(Series.ofBool().isTrue());
+        assertFalse(Series.ofBool().isTrue());
         assertTrue(Series.ofBool(true, true, true).isTrue());
         assertFalse(Series.ofBool(true, false, true).isTrue());
     }